requirements.txt
# Django and Core Dependencies
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11

# Django Extensions
django-mptt==0.18.0
django-cors-headers==4.9.0

# Image Processing
Pillow==12.0.0

# HTTP Requests (for product import)
requests==2.32.5

# Database (optional - for PostgreSQL)
# psycopg2-binary==2.9.11

# Utilities
python-dotenv==1.2.1

# Rich Text Editor for Blog
django-ckeditor==6.7.0

# Tags for Blog Posts
django-taggit==5.0.1

